pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "gitlab-excel-issues"
version = "0.1.0"
description = "A command-line utility for generating GitLab issues from Excel documents."
authors = ["Thomas F. Hagelien <Thomas.F.Hagelien@sintef.no>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.0"
python-gitlab = "^2.10.1"
python-dotenv = "^0.19.2"
openpyxl = "^3.0.9"
colorama = "^0.4.6"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
todo = "gitlab_excel_issues:main"
